Wizard: Fix parameter substitution to work with strings
The regex in sub() was incorrectly matching all non-whitespace
as parameters.  This resulted in a broken header in cfg_cli.h
where the options for prompt (">> ") does have whitespace in it:

(In this case, the unmatched quote causes a compile failure.)

The new regex matches up to the next newline, where the newline
is not preceded by a backslash (\).  This works for strings
with spaces (like ">> ") and also for multi-line #defines.

Signed-off-by: Derek Simkowiak <derek@simkowiak.net>